fix: Decomposition of the SortableLayer. Some layer renaming fixes
